Simplify `_unique_internal` (#2850)
* Simplify `_unique_internal`

Was retrieving the `indices` found by `unique`. However as those are not
used, it makes no sense to get them. Hence we drop requesting that
result from `unique` and simplify the surrounding code in
`_unique_internal`.

* Guarantee dtypes for initial indices and counts

Before we expected these would be constructed as `int64` arrays. However
now we explicitly require `int64` in their construction to ensure that
we get the desired type in both cases.

* Rename `t` to `u` in `_unique_internal`

Makes more sense to use `u` than `t` given it is no longer a `tuple`.

* Enforce `int64` for `arange` in `_unique_internal`

Would be effectively enforced through copying the `arange` result to an
array with `int64` precision. Still it seems better to enforce this when
creating the array with `arange` to begin with.

* Enumerate `r["values"]` in `_unique_internal`

Makes the code a bit more readable by showing more clearly what we are
iterating over are the values.