Minor tweaks to `_unique_internal` optional result handling (#2855)
* Just use `enumerate` in `_unique_internal`

We are not really using any of the fancy features of NumPy's
`ndenumerate`. So we are better off just using `enumerate`.

* Specify `_unique_internal`'s reductions' `out`

Change the array reduction operations to store their outputs to the
specified record in the result array. Should avoid having some extra
copies floating around.

* Skip `for`-loop over values if no extra results

If we are not requiring extra results for the indices or the counts,
there is no reason to even enter a `for`-loop designed to find these
values. So add an `if` before the `for`-loop to block entry unless these
extra results are being computed.

* Find mask for each unique value once

Instead of finding the mask for each unique value once per extra result
being computed, simply compute it once if any extra result should need
it. Then reuse it for both should it be needed.