Test `unique`'s inverse mapping's shape (#2857)
The inverse mapping from `unique` should have a known shape. Also that
shape should be identical to that of the original array (provided to
`unique`) flattened. This makes the inverse mapping different from the
other results that `unique` provides that have an unknown shape. Namely
because the inverse mapping's shape is known a priori as it is not a
function of the original array's contents (though the contents of the
inverse mapping are). Therefore this tweaks `unique`'s tests to ensure
that the shape of the inverse mapping is being checked against the
original data's size.