Add vdot (#2910)
* Add a function for `vdot`

Provides an implementation of NumPy's `vdot` for Dask Arrays by leverage
the Dask Array `dot` implementation and taking the complex conjugate of
the first argument before passing it on to `dot`. Also flattens the
arguments before passing them onto `dot` to follow NumPy's `vdot`'s
behavior.

* Expose `vdot` in Dask Array's public API

* Include a test for vdot

Provides a simple exercise of `vdot` on some test data. In particular
tries a 1-D array and a 2-D array. Compares the results to NumPy. Also
asserts an expectation with respect to commutation of its arguments
(namely the result will be conjugated).

* Document `vdot` in Dask Array's API docs

* Note adding `vdot` in the changelog