Support N-D arrays with matmul (#2909)
* Reimplement matmul methods with the function

Should simplify things a bit while retaining the same functionality.

* Rewrite `matmul` to support N-D arrays

Previously `matmul` was restricted to 1-D and 2-D arrays. This rewrites
`matmul` to support N-D arrays with N > 1. Support for N > 2 is
accomplished by using `atop` with NumPy's `matmul`. Some adjustments to
the arrays are made to handle the 1-D array case correctly and handle
broadcasting in the same way that NumPy's `matmul` would.

* Test `matmul` on N-D arrays

Make sure that `matmul` behaves correctly on a wide range of array
dimensionalities.

* Drop check for unimplemented behavior in `matmul`

Now that we have implemented `matmul` for N-D arrays with arbitrary N,
there is no need to check which case are not implemented as all cases
are implemented.

* Note `matmul`'s N-D support in the changelog