Preserve singleton chunks in fftshift/ifftshift (#2733)
* Rejoin singleton chunks in *fftshift

The `fftshift` and `ifftshift` functions have to split the data into
pieces and reorder those pieces. However the consequence is this can
change the chunking of the data in the final result. Given FFTs can only
operate along dimensions that have one big chunk the length of the
dimension, having chunking like this broken up into pieces by `fftshift`
and `ifftshift` can be surprising to users. To make things a little
easier, make sure that the original chunking is restored in the case of
a single chunk by the end of `fftshift` and `ifftshift` by rechunking.
As we are only rejoining two chunks into one as they were before, this
should not come with any big performance penalties. All other cases that
are split in half and recombined cannot be rechunked without some kind
of significant performance penalties. So we leave those alone. Doing
this should be less surprising to users and make `fftshift` and
`ifftshift` easier to work with.

* Verify *fftshift preserves singleton chunks

Add some tests to make sure that `ifftshift` and `fftshift` rejoin
singleton chunks at the end. This is an expectation of end users. So we
would like to uphold it if possible and ensure we do so in our tests.

* Note new chunk behavior of `fftshift`/`ifftshift`