Test various slicing identities with Dask Arrays (#2978)
Tries a variety of different slicing arguments with Dask Arrays, which
should return the full array (and are known to in advance). Ensures that
in all of these cases the same Dask Array is returned unchanged. This is
done to ensure that Dask Arrays will not add these no-ops to the graph.