Better error on empty bag to_dataframe (#3025)
When no metadata is passed in to `dask.bag.Bag.to_dataframe`, the first
element of the bag is read and used to infer the dtypes. If the first
partition is empty, this would fail with a non-intuitive error. We now
raise a helpful error.

Also added ability to silence warnings in take, and made take warnings
test actually test the warnings.