Minor optimizations to Dask Array's store (#3153)
* Drop end comma from `tuple` in `retrieve_from_ooc`

There was an extra comma at the end of a `tuple` in `retrieve_from_ooc`.
This drops that extraneous comma to improve readability.

* Flatten Array keys passed to optimize in `store`

Make sure that the Dask keys that are passed to `Array`'s
`__dask_optimize__` method are flattened. While the `__dask_optimize__`
method of `Array` does flatten the keys, it is best not to rely on that
in `store` and just ensure they are properly formatted.

* Fuse region/chunk slices when creating store Dask

Instead of leaving the user-provided `region` and chunk slice (i.e.
`index`) to be fused in the `store_chunk` and `load_chunk` functions, go
ahead and fuse them when constructing the Dask graph and leave `region`
as `None`. This should cutdown the work done in `store_chunk` and
`load_chunk`. Also should allow for the `region` argument in both of
these functions to be subsequently dropped.

* Drop region arg from store/load chunk functions

Since `fuse_slice` is now run as part of Dask construction and this
removes the need to pass the `region` on for handling in the store/load
chunk functions, simply drop the `region` arguments and related logic
from `store_chunk` and `load_chunk`.  Makes them a bit easier to follow
and maintain going forward.

* Optimize `store`'s `targets` through `Delayed`

Use the standard way of optimizing Dask Delayed objects to optimize the
targets in advance if needed. Thus if `cull`ing or any other
optimization strategy becomes common place for Delayed objects, it will
be handled through the standard mechanisms instead of us doing it in
some custom way.

As an added bonus, this avoids collect empty dictionaries in the event a
target does not have a Dask (e.g. not Delayed).  Previously `store` had
collected empty dictionaries in this case. So this should be an
improvement when constructing and working with the `SharedDict` object
for `targets` in the common case (i.e. no Delayed targets).

Also provides a check to guarantee we are actually getting Dask Delayed
objects and are not supporting any other Dask collections. If we don't
get a Dask Delayed object, we assume this is some Array-like object.

* Note Delayed target support in store's docstring

For some time, `store` has supported `Delayed` `targets`. The intent
being to support creating a storage location once storage occurs and
possibly handle other things.  However this wasn't noted in the
docstring. So we make that a little clearer by noting `targets` can be
`Delayed`.

* Merge store Dasks together first

Before merging store Dasks with source and target Dasks, go ahead and
merge the store Dasks together alone. Simplifies the merging process a
bit and avoids directly relying on a dependency. Also makes sure that
there is a `SharedDict` for each group of related dictionaries that get
added to final store Dask. Namely the `sources`, `targets`, and now the
`store_chunk`s steps.

* Rename store_dsks_mrg and store_dsks to store_dsk

As we are only really interested in the one Dask graph that represents
all `store_chunk` operations, renamed `store_dsks_mrg` and `store_dsks`
to `store_dsk` to make that clearer.

* Add optional postcompute Dask to retrieve_from_ooc

Provides an option to give `retrieve_from_ooc` a computed Dask as well
to fill in values in the `load_chunk`s part of the Dask. This can be
useful when inlining results from the `persist`ed `store_chunk`s step.
This argument is totally optional and will be used to inline computed
results of `Future`s as the case may be. If this argument is not
provided, then the keys are simply inserted into the new Dask as was the
case before.

* Provide fused `store_chunk`/`load_chunk` strategy

Renames `store_chunk` to `load_store_chunk`. The new `load_store_chunk`
function differs slightly from `store_chunk`'s previous behavior in that
it will skip storing results if `x` is `None`. Also `load_store_chunk`
takes the new `load_stored` option, which allows for reading the results
back in from `out` after storing them. IOW `load_store_chunk` provides
the behavior of `store_chunk`, `load_chunk`, and a fused version of
`store_chunk` and `load_chunk`. By keeping this functionality all within
one function using different options, it should avoid divergence that
could occur with having different functions.

Reimplements `store_chunk` and `load_chunk` to be light wrappers around
`load_store_chunk` with only a few needed arguments. The rest are
handled in some default manner. By doing this, the original behavior of
`store_chunk` and `load_chunk` are restored while simply using
`load_store_chunk` underneath the hood. The `store_chunk` and
`load_chunk` functions are kept to preserve readability in terms of the
code, Dasks generated, and scheduler progress.

Makes some adjustments to `insert_to_ooc` to handle either using
`load_store_chunk` when `return_stored` and `load_stored` are both
`True` or simply using `store_chunk` for all other cases. This results
in some tweaks to the key name used and arguments provided. The purpose
of these adjustments are to ensure `load_store_chunk` shows up when
loading and storing will happen in the same step. Otherwise
`insert_to_ooc` only adds `store_chunk`, which may or may not be
followed by a separate `load_chunk` step depending on whether
`return_stored` is `True` or `False`.

* Rework `store` to inline uses of `store_chunk`s

Refreshes `store` so as to effectively inline all uses of `store_chunk`s
when `return_stored=True`. Thus users only see `load_store_*` nodes
whether or not `compute` is `True` or `False`. This avoids some lock
contention between storing and loading steps when `compute` is `False`.
When `compute` is `True`, this effectively inlines `Future`s or results
that the `store_chunk` step would have returned. Thus making the most
compact Dask possible when `return_stored=True`.

* Change `insert_to_ooc`'s `sto_chunk` to `func`

Appears that `sto_chunk` was more confusing than helpful in the review.
So rename it to `func` to make it clear we intend this to be a different
function and not a typo of `store_chunk`.

* Use dictionary comprehensions in insert/retrieve

In the `insert_to_ooc` and `retrieve_from_ooc` functions, make use of
dictionary comprehensions to simplify the code a bit. These also tend to
be a bit faster than Python `for`-loops.

* Ensure `store` makes depth 1 Dask for compute True

When `return_stored` and `compute` are both set to `True`, ensure that
`store` returns Dask Arrays that contain depth 1 Dasks only. This helps
ensure that the Dask Arrays generated in this case don't have any
lingering connections to the Dask used for storing in the Dask Arrays
returned.

* Check how many times `store` acquires a `Lock`

Verifies how many times a `Lock` instance is acquired in `store` when
using `return_stored`. Namely if `compute=True`, the lock will be
acquired two times for each chunk as storing and loading are separate
steps to allow for storing of chunks to occur asynchronously. However if
`compute=False`, the lock will be acquired only one time for each chunk
as store and load are fused into a single step that acquires the lock
for both parts. This acts as a check to ensure that the `compute=False`
case of `return_stored` does in fact fuse storing and loading into a
single step that holds the lock for both parts.

* Note `store_chunk` inlining in changelog